// src/types.ts
export interface Diagnosis {
	code: string;
	name: string;
	latin?: string;
}

export const Gender = {
	Male: "male",
	Female: "female",
	Other: "other",
} as const;

export type Gender = (typeof Gender)[keyof typeof Gender];

export interface Patient {
	id: string;
	name: string;
	dateOfBirth: string;
	ssn: string;
	gender: Gender;
	occupation: string;
}

export type NonSensitivePatient = Omit<Patient, "ssn">;

export type NewPatient = Omit<Patient, "id">;